# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DHD Livraison HR Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: number
        - name: officeId
          in: query
          schema:
            type: number
        - name: date
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/checkin:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Record employee check-in (requires QR token + GPS coords)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceCheckInput"
      responses:
        "201":
          description: Check-in recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Invalid GPS/QR or outside 5m radius
  /attendance/checkout:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Record employee check-out (requires QR token + GPS coords)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceCheckInput"
      responses:
        "200":
          description: Check-out recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Invalid GPS/QR or outside 5m radius
  /attendance/{id}:
    get:
      operationId: getAttendanceRecord
      tags: [attendance]
      summary: Get attendance record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
    patch:
      operationId: updateAttendanceRecord
      tags: [attendance]
      summary: Admin correction of attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceRecordUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # ── SALARIES ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AttendanceRecord:
      type: object
      required: [id, employeeId, date, createdAt]
      properties:
        id:
          type: number
        employeeId:
          type: number
        employeeName:
          type: ["string", "null"]
        officeId:
          type: number
        officeName:
          type: ["string", "null"]
        date:
          type: string
        checkInTime:
          type: ["string", "null"]
        checkOutTime:
          type: ["string", "null"]
        checkInLat:
          type: ["number", "null"]
        checkInLng:
          type: ["number", "null"]
        checkOutLat:
          type: ["number", "null"]
        checkOutLng:
          type: ["number", "null"]
        workedMinutes:
          type: ["number", "null"]
        lateMinutes:
          type: ["number", "null"]
        overtimeMinutes:
          type: ["number", "null"]
        lateDeduction:
          type: ["number", "null"]
        overtimeBonus:
          type: ["number", "null"]
        isAbsent:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AttendanceCheckInput:
      type: object
      required: [employeeId, qrToken, latitude, longitude]
      properties:
        employeeId:
          type: number
        qrToken:
          type: string
        latitude:
          type: number
        longitude:
          type: number
    AttendanceRecordUpdate:
      type: object
      properties:
        checkInTime:
          type: string
        checkOutTime:
          type: string
        isAbsent:
          type: boolean
        notes:
          type: string

    # SALARIES
...